Omega code complete

+++ fft_asm.h
@@ -2,6 +2,6 @@
 #define FFT_ASM_INCLUDED
 #include <complex.h>
 
-void fft(double complex *X, int N);
+void fft(double complex *X, long N);
 
 #endif // FFT_ASM_INCLUDED
